#256 convert HelpSearch pages to CTF
