+++ src/components/HelpSearch/HelpSearchPropTypes.js
@@ -1,0 +1,5 @@
+import PropTypes from "prop-types";
+
+export const HelpSearchPropTypes = {
+  defaultResultsLang: PropTypes.string.isRequired,
+};
