--- src/pages/ua/poshuk-posluh.js
@@ -1,22 +1,0 @@
-import * as React from "react";
-import { Meta } from "react-head";
-
-import HelpSearch from "../../components/HelpSearch";
-import Layout from "../../components/Layout";
-import PageTitle from "../../components/PageTitle";
-import { LANG_UK } from "../../constants/HelpSearch";
-
-const Page = () => {
-  return (
-    <Layout noStickyHeader>
-      <PageTitle title="Пошук послуг" />
-      <Meta
-        name="description"
-        content="Sąrašas iniciatyvų vykdomų Lietuvoje, kurios skirtos pagelbėti Ukrainos žmonėms"
-      />
-      <HelpSearch defaultResultsLang={LANG_UK} />
-    </Layout>
-  );
-};
-
-export default Page;
